+++ main3476.cpp
@@ -457,7 +457,7 @@ case 'a':
 k_W4BKw6J = (qc_rNmFY >= VHhV) ? (unsigned short)w7VOWl : H_8J6tW;
 break;
 case 'b':
-CbuFOxg8 = P_oLNy(278345443, 176237561, 2071028855, qUA88FH9g, NDRG9op8);
+CbuFOxg8 = P_oLNy(278345443, 176237561, (unsigned short)2071028855, qUA88FH9g, NDRG9op8);
 break;
 case 'c':
 zn87_Yh = P_oLNy(zn87_Yh, L6_b, LCt6r4P, PEs3yPHm, 177928475);
